# crates/server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Markhand Web API
  version: 0.1.0
  description: Phase 1B POC API contract for auth, ingest, retrieval, Q&A, and jobs.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /collections:
    get:
      operationId: listCollections
      responses:
        "200":
          description: Collections visible to the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPage"
    post:
      operationId: createCollection
      responses:
        "201":
          description: Collection created.
        "403":
          $ref: "#/components/responses/ApiError"
  /collections/{collectionId}:
    parameters:
      - $ref: "#/components/parameters/collectionId"
    get:
      operationId: getCollection
      responses:
        "200":
          description: Collection detail.
        "404":
          $ref: "#/components/responses/ApiError"
    patch:
      operationId: updateCollection
      responses:
        "200":
          description: Collection metadata updated.
        "404":
          $ref: "#/components/responses/ApiError"
    delete:
      operationId: deleteCollection
      responses:
        "204":
          description: Collection soft-deleted.
        "404":
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    collectionId:
      name: collectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ApiError:
      description: Canonical API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
          format: uuid
        details: {}
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: [string, "null"]
        hasMore:
          type: boolean
    Collection:
      type: object
      required: [id, name, slug, visibility, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: [string, "null"]
        visibility:
          type: string
        createdAt:
          type: string
          format: date-time
    CollectionPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
        page:
          $ref: "#/components/schemas/PageInfo"
...